Update file paths and add encryption key
The primary action in this commit is the adjustment of several file paths within the project. Most notably, the paths relating to the 'loaderBoot.js' file and scripts in 'buildDSU.js' and 'octopusRun.js' have been updated. Additionally, this commit introduces the "SSO_SECRETS_ENCRYPTION_KEY" to the 'env.json' file. Some extra trailing commas and minor formatting changes are made as well.